GestorJuego y Abstraccion Personaje

+++ core/src/main/java/Gameplay/Movimientos/LanzaRoca.java
@@ -1,15 +1,14 @@
 package Gameplay.Movimientos;
 
-
 import Gameplay.Gestores.GestorProyectiles;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.MathUtils;
-import entidades.Personaje;
-import entidades.Roca;
+import entidades.personajes.Personaje;
+import entidades.proyectiles.Roca;
 
 public class LanzaRoca extends MovimientoRango {
 
-    public LanzaRoca(String nombre, Texture textura, float cooldown, float velocidad, int danio, GestorProyectiles gestorProyectiles) {
+    public LanzaRoca(String nombre, Texture textura, int cooldown, float velocidad, int danio, GestorProyectiles gestorProyectiles) {
         super(nombre, textura, cooldown, velocidad, danio, gestorProyectiles);
     }
