+++ core/src/main/java/Gameplay/Movimientos/MovimientoRango.java
@@ -2,7 +2,7 @@ package Gameplay.Movimientos;
 
 import Gameplay.Gestores.GestorProyectiles;
 import com.badlogic.gdx.graphics.Texture;
-import entidades.Personaje;
+import entidades.personajes.Personaje;
 
 public abstract class MovimientoRango extends Movimiento {
 
@@ -10,7 +10,7 @@ public abstract class MovimientoRango extends Movimiento {
     protected int danio;
     protected GestorProyectiles gestorProyectiles;
 
-    public MovimientoRango(String nombre, Texture textura, float cooldown, float velocidad, int danio, GestorProyectiles gestorProyectiles) {
+    public MovimientoRango(String nombre, Texture textura, int cooldown, float velocidad, int danio, GestorProyectiles gestorProyectiles) {
         super(nombre, textura, cooldown);
         this.velocidad = velocidad;
         this.danio = danio;
